+++ src/main/java/com/codex/composer/internal/data/data/ModRecipeProvider.java
@@ -39,10 +39,7 @@ import net.minecraft.data.recipe.RecipeGenerator;
 //? if minecraft: >=1.21.3
 import net.minecraft.registry.RegistryKeys;
 
-//? legacy {
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
-//? } else
-//import net.fabricmc.fabric.api.datagen.v1.FabricPackOutput;
 
 public class ModRecipeProvider extends FabricRecipeProvider {
     //? if minecraft: <=1.20.4 {
@@ -50,7 +47,7 @@ public class ModRecipeProvider extends FabricRecipeProvider {
         super(output);
     }
     *///? } else {
-    public ModRecipeProvider(/*? if legacy {*/FabricDataOutput/*? } else {*//*FabricPackOutput*//*? }*/ output, CompletableFuture<RegistryWrapper.WrapperLookup> completableFuture) {
+    public ModRecipeProvider(FabricDataOutput output, CompletableFuture<RegistryWrapper.WrapperLookup> completableFuture) {
         super(output, completableFuture);
     }